backend/requirements.txt
flask
Flask-RESTful
python-dotenv
requests
pythonpy
psycopg2-binary 
sqlalchemy 
flask-sqlalchemy